requirements.txt
Flask==3.0.3
google-genai>=1.50.1
python-dotenv==1.0.1

